pyproject.toml
[project]
name = "migratowl"
version = "0.1.0"
description = "AI-powered dependency migration analyzer"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.2.5",
    "fastapi>=0.115.0",
    "langchain-kubernetes[agent-sandbox]>=0.2.0",
    "langchain-anthropic",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv",
    "python-multipart>=0.0.18",
    "httpx>=0.27",
    "packaging>=24.0",
    "html2text>=2024.2.26",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff",
    "langgraph-cli[inmem]",
    "langfuse",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
